fix(davinci): keep custom element ratchets passing
