+++ crates/vize_s1_to_s2/src/lower/cx/custom_element.rs
@@ -1,0 +1,9 @@
+use super::Cx;
+
+impl Cx<'_> {
+    pub(crate) fn is_custom_element(&self, tag: &str) -> bool {
+        self.custom_element_patterns
+            .iter()
+            .any(|pattern| crate::emit::tag_pattern_matches(pattern.as_str(), tag))
+    }
+}
