+++ davinci-road/plan/storage-boundary.md
@@ -22,7 +22,7 @@ or `collections` modules does not bypass the boundary.
 ## Retained `alloc::vec::Vec` inventory
 
 The four library trees in the reviewed inventory contain 78 production files,
-90 direct `alloc::vec::Vec` paths, and 311 bound `Vec`/`StdVec` uses. "Direct"
+90 direct `alloc::vec::Vec` paths, and 312 bound `Vec`/`StdVec` uses. "Direct"
 counts imports and fully-qualified paths; "bound" counts every type,
 constructor, and method path reached through a direct `Vec` import or alias.
 The executable ledger requires strict equality, so both growth and reduction
@@ -32,7 +32,7 @@ must update the file row and aggregate evidence in the same change.
 | -------- | ----: | -----------: | ---------: | ------------------------------------------------------------------------------------------------------------------ |
 | contract |    13 |           24 |         65 | Owned Folio and S2 serialization data has input-defined cardinality and forms a stable contract.                   |
 | analysis |     7 |            8 |         18 | Diagnostics, side tables, filters, and verifier results grow with the input; no inline bound is established.       |
-| lower    |    12 |           12 |         46 | Lowering worklists and owned results grow with source-tree shape. Bounded substructures may migrate independently. |
+| lower    |    12 |           12 |         47 | Lowering worklists and owned results grow with source-tree shape. Bounded substructures may migrate independently. |
 | pass     |    14 |           14 |         55 | Facts, provenance, and traversal worklists grow with the number of operations.                                     |
 | emit     |    32 |           32 |        127 | Ordered output buffers and collected emission inputs grow with the document.                                       |
 
@@ -71,9 +71,9 @@ or count and fails the gate instead of becoming a `no_std` escape from S0.
 | s2       | `vize_s0::String`       |    11 |           11 |         55 |
 | s2       | `vize_s0::Vec`          |     9 |            9 |         17 |
 | s2       | `vize_s0::SmallVec`     |     0 |            0 |          0 |
-| s1_to_s2 | `alloc::vec::Vec`       |    58 |           58 |        228 |
+| s1_to_s2 | `alloc::vec::Vec`       |    58 |           58 |        229 |
 | s1_to_s2 | `alloc::string::String` |     0 |            0 |          0 |
-| s1_to_s2 | `vize_s0::String`       |    84 |           88 |        423 |
+| s1_to_s2 | `vize_s0::String`       |    84 |           88 |        428 |
 | s1_to_s2 | `vize_s0::Vec`          |    15 |           16 |         69 |
 | s1_to_s2 | `vize_s0::SmallVec`     |     4 |            4 |          9 |
